Fix generation of recursive types
An error was being generated when a type included itself. For example:

struct list {
    int val;
    list *next;
}

In the generation phase, the 'list' typedef would recurse before caching a
value. This meant that when *next was resolved, it would resolve to the top
level typedef and call generate() again, rather than simply link to the existing
value.

This is resolved by splitting generation into 2 phases. The first generates the
target class, but doesn't initialize it, therefore doesn't need to recurse. The
output of this first phase is cached. The second phase does initialization, and
can recurse. This means that in the above case, *next can resolve to the top
level typedef, even though it hasn't been initialized yet. It will be
initialized when the top level typedef completes.

This change also makes the resolution of typerefs iterative, which further
simplifies recursion.